+++ src/sentence/Sentence.spec.ts
@@ -1,5 +1,6 @@
 import { mockRandom, resetMockRandom } from 'jest-mock-random';
-import { Sentence, getTotalWeightOfEntries } from './Sentence';
+import { Sentence } from './Sentence';
+import { getTotalWeightOfEntries } from './utils/array';
 
 describe('Sentence.js', () => {
   const template= 'Let\'s {verb} this, and hope for the {adjective}.';
@@ -221,6 +222,12 @@ describe('Sentence.js', () => {
     describe('placeholderNotation', () => {
       it('should default to curly brackets', () => {
         const { value } = new Sentence(helloWorldTemplate, helloWorldVocab);
+        expect(value).toEqual('Hello, world.');
+      });
+
+      it('should be able to use special characters', () => {
+        const bracketsTemplate = '[greeting], [noun].';
+        const { value } = new Sentence(bracketsTemplate, helloWorldVocab, { placeholderNotation: '[ ]' });
         expect(value).toEqual('Hello, world.');
       });
     });